docker/requirements.txt
fastapi==0.75.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.36
sqlalchemy-utils==0.38.2
uvicorn==0.17.6
